backend/requirements.txt
alembic==1.7.5
aniso8601==9.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coverage==6.2
Flask==2.0.2
Flask-Classy==0.6.10
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
flask-swagger-ui==3.36.0
google-api-core==2.3.1
google-auth==2.3.3
google-cloud-core==2.2.1
googleapis-common-protos==1.54.0
greenlet==1.1.2
gunicorn==20.1.0
haversine==2.5.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
protobuf==3.19.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.4.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.28
urllib3==1.26.7
Werkzeug==2.0.2
